+++ processor/logtypedetectionprocessor/fingerprint/fingerprint_json.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package logtypedetectionprocessor
+package fingerprint
 
 var jsonStructuralChars [256]bool
