+++ processor/logtypedetectionprocessor/fingerprint/fingerprint_xml.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package logtypedetectionprocessor
+package fingerprint
 
 import "strings"
